Pruebas/Backend/FastApi/requirements.txt
fastapi
python-jose
passlib
bcrypt
pymongo
python-multipart
uvicorn
routers